(vm) rename classes and packages #103
