+++ vm/src/main/scala/pravda/vm/ExecutionResult.scala
@@ -1,0 +1,5 @@
+package pravda.vm
+
+final case class ExecutionResult(memory: Memory, error: Option[VmErrorException], wattCounter: WattCounter) {
+  def isSuccess: Boolean = error.isEmpty
+}
