+++ vm/src/main/scala/pravda/vm/ProgramContext.scala
@@ -1,4 +1,4 @@
-package pravda.vm.state
+package pravda.vm
 
 import java.nio.ByteBuffer
