+++ vm/src/main/scala/pravda/vm/VmError.scala
@@ -1,6 +1,4 @@
-package pravda.vm.state
-
-import pravda.common.domain.Address
+package pravda.vm
 
 sealed abstract class VmError(val code: Int)
 
@@ -24,29 +22,4 @@ object VmError {
   case object ExtCallStackOverflow extends VmError(601)
 
   final case class SomethingWrong(ex: Throwable) extends VmError(999)
-
 }
-
-final case class VmErrorException(error: VmError, stackTrace: StackTrace = StackTrace.empty) extends Exception {
-  def addToTrace(p: Point): VmErrorException = copy(stackTrace = stackTrace + p)
-}
-
-final case class StackTrace(stackTrace: Seq[Point]) {
-
-  def +(p: Point): StackTrace = {
-    copy(stackTrace = p +: stackTrace)
-  }
-}
-
-object StackTrace {
-
-  def apply(p: Point): StackTrace = {
-    StackTrace(List(p))
-  }
-
-  def empty: StackTrace = {
-    StackTrace(List.empty[Point])
-  }
-}
-
-final case class Point(callStack: Seq[Int], position: Int, address: Option[Address])
