+++ vm/src/main/scala/pravda/vm/operations/NativeCoinOperations.scala
@@ -1,10 +1,9 @@
 package pravda.vm.operations
 
 import pravda.common.domain
-import pravda.vm.state.VmError.OperationDenied
-import pravda.vm.state.{Environment, Memory, VmErrorException}
-import pravda.vm.watt.WattCounter
-import pravda.vm.watt.WattCounter._
+import pravda.vm.VmError.OperationDenied
+import pravda.vm.WattCounter._
+import pravda.vm.{Environment, Memory, VmErrorException, WattCounter}
 
 final class NativeCoinOperations(memory: Memory,
                                  environment: Environment,
@@ -14,14 +13,14 @@ final class NativeCoinOperations(memory: Memory,
   def ptransfer(): Unit = maybeProgramAddress match {
     case None => throw VmErrorException(OperationDenied)
     case Some(programAddress) =>
-      val amount = coin(memory.pop())
+      val amount = coins(memory.pop())
       val to = address(memory.pop())
       wattCounter.cpuUsage(CpuStorageUse)
       environment.transfer(programAddress, to, amount)
   }
 
   def transfer(): Unit = {
-    val amount = coin(memory.pop())
+    val amount = coins(memory.pop())
     val to = address(memory.pop())
     wattCounter.cpuUsage(CpuStorageUse)
     environment.transfer(environment.executor, to, amount)
